pyproject.toml
[project]
name = "substack-archives-downloader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.32.5",
    "selenium>=4.0.0",
    "soupsieve==2.2.1",
    "urllib3>=2.6.1",
    "validators==0.18.2",
    "webdriver-manager>=4.0.2",
    "python-dotenv>=1.0.0",
]
